solomonresearch/blockchain sna litrev/RIS files/literature_review_blockchain_sna/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
anthropic>=0.18.0
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6
jinja2>=3.1.2
pyyaml>=6.0
requests>=2.31.0
tenacity>=8.2.0